+++ Utilities/openjpeg/mqc.c
@@ -355,7 +355,7 @@ void mqc_destroy(opj_mqc_t *mqc) {
 }
 
 int mqc_numbytes(opj_mqc_t *mqc) {
-  return mqc->bp - mqc->start;
+  return (int)(mqc->bp - mqc->start);
 }
 
 void mqc_init_enc(opj_mqc_t *mqc, unsigned char *bp) {
